refactor: 工程化歌词获取相关接口
