+++ src/renderer/api/lyrics/lrcs/normal/kugou.ts
