+++ src/renderer/api/lyrics/lrcs/normal/lrclib.ts
