+++ src/renderer/api/lyrics/lrcs/normal/netease.ts
