+++ src/renderer/api/lyrics/lrcs/normal/sodaMusic.ts
