+++ src/renderer/api/lyrics/lrcs/normal/types.ts
